+++ src/docs/asciidoc/member.adoc
@@ -76,7 +76,9 @@ include::{snippets}/MemberApi/Register/Process/request-fields.adoc[]
 include::{snippets}/MemberApi/Register/Process/http-response.adoc[]
 include::{snippets}/MemberApi/Register/Process/response-fields.adoc[]
 
-== 마이페이지 정보 수정 (AccessToken)
+== 마이페이지
+
+=== 내 정보 수정 (AccessToken)
 
 > HTTP Request
